add STL

+++ build_occt.sh
@@ -1,16 +1,18 @@
 
-curl  -L -o occt7.1.0.tgz "http://git.dev.opencascade.org/gitweb/?p=occt.git;a=snapshot;h=89aebdea8d6f4d15cfc50e9458cd8e2e25022326;sf=tgz"
 
-tar -xf occt7.1.0.tgz
+if [! -f occt7.1.0.tgz ]
+then 
+  curl  -L -o occt7.1.0.tgz "http://git.dev.opencascade.org/gitweb/?p=occt.git;a=snapshot;h=89aebdea8d6f4d15cfc50e9458cd8e2e25022326;sf=tgz"
+  tar -xf occt7.1.0.tgz
+  mv occt-89aebde occt-7.1.0
 
-mv occt-89aebde occt-7.1.0
-
-echo -----------------------------------------------------------------
-echo          PATCHING 7.1.0 TO SPEEDUP BUILD
-echo -----------------------------------------------------------------
-cd occt-7.1.0
-patch -p1 < ../add_cotire_to_7.1.0.patch
-cd ..
+  echo -----------------------------------------------------------------
+  echo          PATCHING 7.1.0 TO SPEEDUP BUILD
+  echo -----------------------------------------------------------------
+  cd occt-7.1.0
+  patch -p1 < ../add_cotire_to_7.1.0.patch
+  cd ..
+fi
 
 mkdir build
 cd build
